Added NPC to unlock abilities
Abilities are unlocked by running past the NPC.
Altered Melee Attack to happen when Z is pressed, as per the design doc.
Tried to fix sawed off but running into issue where multiple bullets are spawned if you spam the dash button.
Edited basic Enemy health to be in line with design doc.